+++ tools/windows/MSStoreApps/MSStoreApps.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.14.36623.8
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "MSStoreApps", "MSStoreApps\MSStoreApps.vcxproj", "{A8D4744D-505B-42F3-B56F-27186A32A6CF}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{A8D4744D-505B-42F3-B56F-27186A32A6CF}.Debug|x64.ActiveCfg = Release|x64
+		{A8D4744D-505B-42F3-B56F-27186A32A6CF}.Debug|x64.Build.0 = Release|x64
+		{A8D4744D-505B-42F3-B56F-27186A32A6CF}.Debug|x86.ActiveCfg = Debug|Win32
+		{A8D4744D-505B-42F3-B56F-27186A32A6CF}.Debug|x86.Build.0 = Debug|Win32
+		{A8D4744D-505B-42F3-B56F-27186A32A6CF}.Release|x64.ActiveCfg = Release|x64
+		{A8D4744D-505B-42F3-B56F-27186A32A6CF}.Release|x64.Build.0 = Release|x64
+		{A8D4744D-505B-42F3-B56F-27186A32A6CF}.Release|x86.ActiveCfg = Release|Win32
+		{A8D4744D-505B-42F3-B56F-27186A32A6CF}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {E30F5E45-0208-4485-8F1F-2456140CF60E}
+	EndGlobalSection
+EndGlobal
